Relocate umi to /lib
